MedAffairs_GenAI/main.py
#Loading Libraries(te=4.2.0)
from langchain.text_splitter import CharacterTextSplitter
from langchain.document_loaders import UnstructuredFileLoader
from langchain.vectorstores import FAISS
from langchain.embeddings import OpenAIEmbeddings
from langchain.chains import LLMChain
from langchain.prompts import PromptTemplate
from langchain.chat_models import ChatOpenAI
from langchain.chains import RetrievalQA
import tiktoken
import openai
import spacy
import os
import pandas as pd
import config
from numpy import ceil
from PyPDF2 import PdfReader
from langchain.chains.summarize import load_summarize_chain
import docx
import io
import re
import docx
from docx import Document


def print_hi(name):
    # Use a breakpoint in the code line below to debug your script.
    print(f'Hi, {name}')  # Press Ctrl+F8 to toggle the breakpoint.


# Press the green button in the gutter to run the script.
if __name__ == '__main__':
    print_hi('PyCharm')

# See PyCharm help at https://www.jetbrains.com/help/pycharm/
