MedAffairs_GenAI/requirements.txt
python-docx
pandas
tiktoken
langchain
numpy
rouge-score
nltk
openai
pdfplumber
flask
